# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Control your legacy Reciva based internet radios (Crane, Grace Digital, Tangent, etc.) via web browser or REST API.",
        "title": "Reciva Web Remote",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
        "/build": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "build"
                ],
                "summary": "Get build",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Build"
                        }
                    }
                }
            }
        },
        "/presets": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presets"
                ],
                "summary": "List presets",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Preset"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presets"
                ],
                "summary": "Update preset",
                "parameters": [
                    {
                        "description": "Preset",
                        "name": "preset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Preset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            }
        },
        "/presets/{url}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "presets"
                ],
                "summary": "Get preset",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Preset URL",
                        "name": "url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Preset"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            }
        },
        "/radios": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "radios"
                ],
                "summary": "List radios",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Radio"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "radios"
                ],
                "summary": "Discover radios",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "409": {
                        "description": "Discovery already in progress",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            }
        },
        "/radios/{uuid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "radios"
                ],
                "summary": "Get radio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Radio UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Radio"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "radios"
                ],
                "summary": "Delete radio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Radio UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            }
        },
        "/radios/{uuid}/subscription": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "radios"
                ],
                "summary": "Refresh radio subscription",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Radio UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            }
        },
        "/radios/{uuid}/volume": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "radios"
                ],
                "summary": "Refresh radio volume",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Radio UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            }
        },
        "/states": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "List states",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/state.State"
                            }
                        }
                    }
                }
            }
        },
        "/states/{uuid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "states"
                ],
                "summary": "Get state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Radio UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/state.State"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "states"
                ],
                "summary": "Update state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Radio UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Patch state",
                        "name": "state",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.PostState"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "http.HTTPError": {
            "type": "object",
            "required": [
                "message"
            ],
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "http.PostState": {
            "type": "object",
            "properties": {
                "audio_source": {
                    "type": "string"
                },
                "power": {
                    "type": "boolean"
                },
                "preset": {
                    "type": "integer"
                },
                "volume": {
                    "type": "integer"
                },
                "volume_delta": {
                    "type": "integer"
                }
            }
        },
        "model.Build": {
            "type": "object",
            "required": [
                "built_by",
                "commit",
                "date",
                "release_url",
                "summary",
                "version"
            ],
            "properties": {
                "built_by": {
                    "type": "string"
                },
                "commit": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "release_url": {
                    "type": "string"
                },
                "summary": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "model.Preset": {
            "type": "object",
            "required": [
                "title_new",
                "url",
                "url_new"
            ],
            "properties": {
                "title_new": {
                    "description": "TitleNew is the overridden title.",
                    "type": "string"
                },
                "url": {
                    "description": "URL of the preset.",
                    "type": "string"
                },
                "url_new": {
                    "description": "URLNew is the overridden URL.",
                    "type": "string"
                }
            }
        },
        "model.Radio": {
            "type": "object",
            "required": [
                "name",
                "uuid"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "pubsub.Stale": {
            "type": "string",
            "enum": [
                "radios",
                "presets"
            ],
            "x-enum-varnames": [
                "StaleRadios",
                "StalePresets"
            ]
        },
        "pubsub.Topic": {
            "type": "string",
            "enum": [
                "state",
                "stale.state.hook",
                "discover",
                "stale.radios",
                "error",
                "stale"
            ],
            "x-enum-varnames": [
                "TopicState",
                "TopicStaleStateHook",
                "TopicDiscover",
                "TopicStaleRadios",
                "TopicError",
                "TopicStale"
            ]
        },
        "state.Preset": {
            "type": "object",
            "required": [
                "number",
                "title",
                "title_new",
                "url",
                "url_new"
            ],
            "properties": {
                "number": {
                    "description": "Number is the preset number.",
                    "type": "integer"
                },
                "title": {
                    "description": "Title of the preset.",
                    "type": "string"
                },
                "title_new": {
                    "description": "TitleNew is the overridden title.",
                    "type": "string"
                },
                "url": {
                    "description": "URL of the preset.",
                    "type": "string"
                },
                "url_new": {
                    "description": "URLNew is the overridden URL.",
                    "type": "string"
                }
            }
        },
        "state.State": {
            "type": "object",
            "required": [
                "audio_source",
                "audio_sources",
                "is_muted",
                "metadata",
                "model_name",
                "model_number",
                "name",
                "power",
                "preset_number",
                "presets",
                "status",
                "title",
                "title_new",
                "url",
                "url_new",
                "uuid",
                "volume"
            ],
            "properties": {
                "audio_source": {
                    "description": "AudioSource is the audio source.",
                    "type": "string"
                },
                "audio_sources": {
                    "description": "AudioSources is the list of available audio sources.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "is_muted": {
                    "description": "IsMuted represents if the radio is muted.",
                    "type": "boolean"
                },
                "metadata": {
                    "description": "Metadata of the current playing stream.",
                    "type": "string"
                },
                "model_name": {
                    "description": "ModelName is the model name of the device.",
                    "type": "string"
                },
                "model_number": {
                    "description": "ModelNumber is the model number of the device.",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the radio.",
                    "type": "string"
                },
                "power": {
                    "description": "Power represents if the radio is not in standby.",
                    "type": "boolean"
                },
                "preset_number": {
                    "description": "PresetNumber is the current preset that is playing.",
                    "type": "integer"
                },
                "presets": {
                    "description": "Presets of the radio.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/state.Preset"
                    }
                },
                "status": {
                    "description": "Status is either playing, connecting, or stopped.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/state.Status"
                        }
                    ]
                },
                "title": {
                    "description": "Title of the current playing stream.",
                    "type": "string"
                },
                "title_new": {
                    "description": "TitleNew is the overridden title.",
                    "type": "string"
                },
                "url": {
                    "description": "URL of the stream that is currently selected.",
                    "type": "string"
                },
                "url_new": {
                    "description": "URLNew is the overridden URL.",
                    "type": "string"
                },
                "uuid": {
                    "description": "UUID of the radio.",
                    "type": "string"
                },
                "volume": {
                    "description": "Volume of the radio.",
                    "type": "integer"
                }
            }
        },
        "state.Status": {
            "type": "string",
            "enum": [
                "Connecting",
                "Playing",
                "Stopped",
                ""
            ],
            "x-enum-varnames": [
                "StatusConnecting",
                "StatusPlaying",
                "StatusStopped",
                "StatusUnknown"
            ]
        },
        "ws.Command": {
            "type": "object",
            "properties": {
                "state": {
                    "$ref": "#/definitions/ws.CommandState"
                },
                "subscribe": {
                    "$ref": "#/definitions/ws.CommandSubscribe"
                }
            }
        },
        "ws.CommandState": {
            "type": "object",
            "required": [
                "partial",
                "uuid"
            ],
            "properties": {
                "partial": {
                    "type": "boolean"
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "ws.CommandSubscribe": {
            "type": "object",
            "required": [
                "topics"
            ],
            "properties": {
                "topics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pubsub.Topic"
                    }
                }
            }
        },
        "ws.Event": {
            "type": "object",
            "required": [
                "data",
                "topic"
            ],
            "properties": {
                "data": {},
                "topic": {
                    "$ref": "#/definitions/pubsub.Topic"
                }
            }
        }
    }
}